push all the random stuff in my dir
