+++ frontend-1/postcss.config.cjs
@@ -1,4 +1,6 @@
 module.exports = {
+  mode: 'jit',
+  purge: ['./src/**/*.{js,jsx,ts,tsx'],
   plugins: {
     '@tailwindcss/postcss': {}, // Explicitly use the new package
     autoprefixer: {}
