+++ frontend-1/src/lib/components/Sidebar.svelte
@@ -1,0 +1,13 @@
+    <div class="bg-gray-800 w-52 text-white flex flex-col items-center space-y-4 relative">
+        <h2 class="flex flex-col items-center pt-5 font-bold text-xl">OpenLabs</h2>
+        <div class="flex flex-col items-center pt-0">
+            <img class="w-32 mx-auto" src="https://avatars.githubusercontent.com/u/196604745?s=200&v=4" alt="OpenLabs Logo" />
+        </div>
+        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full w-7/8">Deployed Ranges</button>
+        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full w-7/8">Range Templates</button>
+        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full w-7/8">Documentation</button>
+        <div class="flex items-center p-4 absolute bottom-0 w-full flex-col items-center">
+            <img class="w-12 h-12 rounded-full mb-2" src="https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png" alt="User Avatar" />
+            <p>John Doe</p>
+        </div>
+    </div>
